requirements.txt
django
react
djangorestframework
python-decouple